Added support for Eshell
Eshell now has a (use-package) declaration and some helper defuns which make it
more useful.

Bindings have been set to open or close eshell on C-!.
A global variable my:eshell-visual-commands holds a list of names of all the
programs which will need to be run in an ansi-term.

dot-setup has been modified to support an eshell/alias file. This will install
to ~/.emacs.d/eshell/alias.